🎨 Move vtk utility
